refactor(codex): split login runner from account manager (1k-line bar)
